+++ components/animations/effectTimings/types/composite.js
