+++ components/animations/effectTimings/types/delay.js
