+++ components/animations/effectTimings/types/direction.js
