+++ components/animations/effectTimings/types/duration.js
