+++ components/animations/effectTimings/types/fill.js
