+++ components/animations/effectTimings/types/index.js
