+++ components/animations/effectTimings/types/iterationComposite.js
