+++ components/animations/effectTimings/types/pseudoElement.js
