--- components/customRadios/effectTimingRadio.js
@@ -1,5 +1,0 @@
-import AbstractRadio from "./abstractRadio";
-
-export default function EffectTimingRadio() {
-
-}
