+++ components/customRadios/index.js
@@ -1,3 +1,3 @@
-export {default as EffectTimingRadio} from './effectTimingRadio'
+
 export {default as UpdatePlaybackRateRadio} from './updatePlaybackrateRadio'
 export {default as FillRadio} from './fillRadio'
