+++ components/customTabs/animationApiTab.js
@@ -1,7 +1,7 @@
 import AbstractTab from "./abstractTab";
 
 export default function AnimationApiTab({set}) {
-    const tabs = ["properties", "effect-timing", "methods", "steps"];
+    const tabs = ["effect-timing", "methods"];
     const animeStyle = {
         background: "none",
         borderBottom:'0.2rem solid #00edff',
