+++ lib/theme.js
@@ -27,16 +27,6 @@ const styles = {
   ) 
 }
 
-//   const theme = extendTheme({ breakpoints,fonts: {
-//     logo: `'Times New Roman', Times, sans-serif`,   
-//   }, 
-//     config: {
-//       initialColorMode: 'dark',
-//       useSystemColorMode: false,
-//     }
-//   });
-
-
 // 2. Add your color mode config
 const config = {
   initialColorMode: 'dark',
